+++ pallet/ethtx-forwarder/src/tests/eip2930.rs
@@ -18,7 +18,7 @@
 
 // darwinia
 use crate::{mock::*, tests::*, RuntimeEthOrigin};
-use ethereum::EIP1559Transaction;
+use ethereum::EIP2930Transaction;
 // frontier
 use fp_evm::FeeCalculator;
 // substrate
@@ -26,11 +26,10 @@ use frame_support::{assert_err, assert_ok, traits::Currency};
 use sp_core::U256;
 use sp_runtime::{DispatchError, ModuleError};
 
-pub fn eip1559_erc20_creation_unsigned_transaction() -> EIP1559UnsignedTransaction {
-	EIP1559UnsignedTransaction {
+fn eip2930_erc20_creation_unsigned_transaction() -> EIP2930UnsignedTransaction {
+	EIP2930UnsignedTransaction {
 		nonce: U256::zero(),
-		max_priority_fee_per_gas: U256::from(1),
-		max_fee_per_gas: U256::from(1),
+		gas_price: U256::from(1),
 		gas_limit: U256::from(1_000_000),
 		action: ethereum::TransactionAction::Create,
 		value: U256::zero(),
@@ -39,16 +38,17 @@ pub fn eip1559_erc20_creation_unsigned_transaction() -> EIP1559UnsignedTransacti
 }
 
 #[test]
-fn test_eip1559_transaction_works() {
+fn test_eip2930_transaction_works() {
 	let alice = address_build(1);
 
 	ExtBuilder::default()
 		.with_balances(vec![(alice.address, 1_000_000_000_000)])
 		.build()
 		.execute_with(|| {
-			let unsigned_tx = eip1559_erc20_creation_unsigned_transaction();
+			let unsigned_tx = eip2930_erc20_creation_unsigned_transaction();
 			let t = unsigned_tx.sign(&alice.private_key, None);
-			assert_ok!(RuntimeTransact::runtime_transact(
+
+			assert_ok!(EthTxForwarder::runtime_transact(
 				RuntimeEthOrigin::RuntimeTransact(alice.address).into(),
 				Box::new(t)
 			));
@@ -59,18 +59,18 @@ fn test_eip1559_transaction_works() {
 }
 
 #[test]
-fn test_eip1559_transaction_with_auto_nonce() {
+fn test_eip2930_transaction_with_auto_nonce() {
 	let alice = address_build(1);
 
 	ExtBuilder::default()
 		.with_balances(vec![(alice.address, 1_000_000_000_000)])
 		.build()
 		.execute_with(|| {
-			let mut unsigned_tx = eip1559_erc20_creation_unsigned_transaction();
+			let mut unsigned_tx = eip2930_erc20_creation_unsigned_transaction();
 			unsigned_tx.nonce = U256::MAX;
 			let t = unsigned_tx.sign(&alice.private_key, None);
 
-			assert_ok!(RuntimeTransact::runtime_transact(
+			assert_ok!(EthTxForwarder::runtime_transact(
 				RuntimeEthOrigin::RuntimeTransact(alice.address).into(),
 				Box::new(t)
 			));
@@ -81,19 +81,18 @@ fn test_eip1559_transaction_with_auto_nonce() {
 }
 
 #[test]
-fn test_eip1559_transaction_with_auto_gas_price() {
+fn test_eip2930_transaction_with_auto_gas_price() {
 	let alice = address_build(1);
 
 	ExtBuilder::default()
 		.with_balances(vec![(alice.address, 1_000_000_000_000)])
 		.build()
 		.execute_with(|| {
-			let mut unsigned_tx = eip1559_erc20_creation_unsigned_transaction();
-			unsigned_tx.max_fee_per_gas =
+			let mut unsigned_tx = eip2930_erc20_creation_unsigned_transaction();
+			unsigned_tx.gas_price =
 				<Runtime as pallet_evm::Config>::FeeCalculator::min_gas_price().0 - 1;
 			let t = unsigned_tx.sign(&alice.private_key, None);
-
-			assert_ok!(RuntimeTransact::runtime_transact(
+			assert_ok!(EthTxForwarder::runtime_transact(
 				RuntimeEthOrigin::RuntimeTransact(alice.address).into(),
 				Box::new(t)
 			));
@@ -104,14 +103,14 @@ fn test_eip1559_transaction_with_auto_gas_price() {
 }
 
 #[test]
-fn test_eip1559_transaction_with_sufficient_balance() {
+fn test_eip2930_transaction_with_sufficient_balance() {
 	let alice = address_build(1);
 	ExtBuilder::default().build().execute_with(|| {
-		let unsigned_tx = eip1559_erc20_creation_unsigned_transaction();
+		let unsigned_tx = eip2930_erc20_creation_unsigned_transaction();
 		let t = unsigned_tx.sign(&alice.private_key, None);
 
 		assert_err!(
-			RuntimeTransact::runtime_transact(
+			EthTxForwarder::runtime_transact(
 				RuntimeEthOrigin::RuntimeTransact(alice.address).into(),
 				Box::new(t.clone())
 			),
@@ -122,9 +121,9 @@ fn test_eip1559_transaction_with_sufficient_balance() {
 			})
 		);
 
-		let fee = RuntimeTransact::total_payment((&t).into());
+		let fee = EthTxForwarder::total_payment((&t).into());
 		let _ = Balances::deposit_creating(&alice.address, fee.as_u64());
-		assert_ok!(RuntimeTransact::runtime_transact(
+		assert_ok!(EthTxForwarder::runtime_transact(
 			RuntimeEthOrigin::RuntimeTransact(alice.address).into(),
 			Box::new(t)
 		));
@@ -135,17 +134,16 @@ fn test_eip1559_transaction_with_sufficient_balance() {
 }
 
 #[test]
-fn test_eip1559_transaction_with_valid_signature() {
+fn test_eip2930_transaction_with_valid_signature() {
 	let alice = address_build(1);
 	ExtBuilder::default()
 		.with_balances(vec![(alice.address, 1_000_000_000_000)])
 		.build()
 		.execute_with(|| {
-			let t = EIP1559Transaction {
+			let t = EIP2930Transaction {
 				chain_id: 0,
 				nonce: U256::zero(),
-				max_priority_fee_per_gas: U256::zero(),
-				max_fee_per_gas: U256::zero(),
+				gas_price: U256::from(1),
 				gas_limit: U256::from(1_000_000),
 				action: ethereum::TransactionAction::Create,
 				value: U256::zero(),
@@ -165,9 +163,9 @@ fn test_eip1559_transaction_with_valid_signature() {
 					213, 144, 24, 205, 4, 67, 173, 192, 144, 53, 144, 193, 107, 2, 176,
 				]),
 			};
-			assert_ok!(RuntimeTransact::runtime_transact(
+			assert_ok!(EthTxForwarder::runtime_transact(
 				RuntimeEthOrigin::RuntimeTransact(alice.address).into(),
-				Box::new(Transaction::EIP1559(t))
+				Box::new(Transaction::EIP2930(t))
 			));
 
 			assert!(System::events()
