+++ pallet/ethtx-forwarder/src/tests/legacy.rs
@@ -18,7 +18,7 @@
 
 // darwinia
 use crate::{mock::*, tests::*, RuntimeEthOrigin};
-use ethereum::EIP2930Transaction;
+use ethereum::LegacyTransaction;
 // frontier
 use fp_evm::FeeCalculator;
 // substrate
@@ -26,8 +26,8 @@ use frame_support::{assert_err, assert_ok, traits::Currency};
 use sp_core::U256;
 use sp_runtime::{DispatchError, ModuleError};
 
-fn eip2930_erc20_creation_unsigned_transaction() -> EIP2930UnsignedTransaction {
-	EIP2930UnsignedTransaction {
+pub fn legacy_erc20_creation_unsigned_transaction() -> LegacyUnsignedTransaction {
+	LegacyUnsignedTransaction {
 		nonce: U256::zero(),
 		gas_price: U256::from(1),
 		gas_limit: U256::from(1_000_000),
@@ -38,17 +38,15 @@ fn eip2930_erc20_creation_unsigned_transaction() -> EIP2930UnsignedTransaction {
 }
 
 #[test]
-fn test_eip2930_transaction_works() {
+fn test_legacy_transaction_works() {
 	let alice = address_build(1);
-
 	ExtBuilder::default()
 		.with_balances(vec![(alice.address, 1_000_000_000_000)])
 		.build()
 		.execute_with(|| {
-			let unsigned_tx = eip2930_erc20_creation_unsigned_transaction();
-			let t = unsigned_tx.sign(&alice.private_key, None);
+			let t = legacy_erc20_creation_unsigned_transaction().sign(&alice.private_key);
 
-			assert_ok!(RuntimeTransact::runtime_transact(
+			assert_ok!(EthTxForwarder::runtime_transact(
 				RuntimeEthOrigin::RuntimeTransact(alice.address).into(),
 				Box::new(t)
 			));
@@ -59,18 +57,17 @@ fn test_eip2930_transaction_works() {
 }
 
 #[test]
-fn test_eip2930_transaction_with_auto_nonce() {
+fn test_legacy_transaction_with_auto_nonce() {
 	let alice = address_build(1);
-
 	ExtBuilder::default()
 		.with_balances(vec![(alice.address, 1_000_000_000_000)])
 		.build()
 		.execute_with(|| {
-			let mut unsigned_tx = eip2930_erc20_creation_unsigned_transaction();
+			let mut unsigned_tx = legacy_erc20_creation_unsigned_transaction();
 			unsigned_tx.nonce = U256::MAX;
-			let t = unsigned_tx.sign(&alice.private_key, None);
+			let t = unsigned_tx.sign(&alice.private_key);
 
-			assert_ok!(RuntimeTransact::runtime_transact(
+			assert_ok!(EthTxForwarder::runtime_transact(
 				RuntimeEthOrigin::RuntimeTransact(alice.address).into(),
 				Box::new(t)
 			));
@@ -81,18 +78,18 @@ fn test_eip2930_transaction_with_auto_nonce() {
 }
 
 #[test]
-fn test_eip2930_transaction_with_auto_gas_price() {
+fn test_legacy_transaction_with_auto_gas_price() {
 	let alice = address_build(1);
-
 	ExtBuilder::default()
 		.with_balances(vec![(alice.address, 1_000_000_000_000)])
 		.build()
 		.execute_with(|| {
-			let mut unsigned_tx = eip2930_erc20_creation_unsigned_transaction();
+			let mut unsigned_tx = legacy_erc20_creation_unsigned_transaction();
 			unsigned_tx.gas_price =
 				<Runtime as pallet_evm::Config>::FeeCalculator::min_gas_price().0 - 1;
-			let t = unsigned_tx.sign(&alice.private_key, None);
-			assert_ok!(RuntimeTransact::runtime_transact(
+			let t = unsigned_tx.sign(&alice.private_key);
+
+			assert_ok!(EthTxForwarder::runtime_transact(
 				RuntimeEthOrigin::RuntimeTransact(alice.address).into(),
 				Box::new(t)
 			));
@@ -103,14 +100,12 @@ fn test_eip2930_transaction_with_auto_gas_price() {
 }
 
 #[test]
-fn test_eip2930_transaction_with_sufficient_balance() {
+fn test_legacy_transaction_with_sufficient_balance() {
 	let alice = address_build(1);
 	ExtBuilder::default().build().execute_with(|| {
-		let unsigned_tx = eip2930_erc20_creation_unsigned_transaction();
-		let t = unsigned_tx.sign(&alice.private_key, None);
-
+		let t = legacy_erc20_creation_unsigned_transaction().sign(&alice.private_key);
 		assert_err!(
-			RuntimeTransact::runtime_transact(
+			EthTxForwarder::runtime_transact(
 				RuntimeEthOrigin::RuntimeTransact(alice.address).into(),
 				Box::new(t.clone())
 			),
@@ -121,9 +116,9 @@ fn test_eip2930_transaction_with_sufficient_balance() {
 			})
 		);
 
-		let fee = RuntimeTransact::total_payment((&t).into());
+		let fee = EthTxForwarder::total_payment((&t).into());
 		let _ = Balances::deposit_creating(&alice.address, fee.as_u64());
-		assert_ok!(RuntimeTransact::runtime_transact(
+		assert_ok!(EthTxForwarder::runtime_transact(
 			RuntimeEthOrigin::RuntimeTransact(alice.address).into(),
 			Box::new(t)
 		));
@@ -134,38 +129,35 @@ fn test_eip2930_transaction_with_sufficient_balance() {
 }
 
 #[test]
-fn test_eip2930_transaction_with_valid_signature() {
+fn test_legacy_transaction_with_valid_signature() {
 	let alice = address_build(1);
 	ExtBuilder::default()
 		.with_balances(vec![(alice.address, 1_000_000_000_000)])
 		.build()
 		.execute_with(|| {
-			let t = EIP2930Transaction {
-				chain_id: 0,
+			let t = LegacyTransaction {
 				nonce: U256::zero(),
 				gas_price: U256::from(1),
 				gas_limit: U256::from(1_000_000),
 				action: ethereum::TransactionAction::Create,
 				value: U256::zero(),
 				input: array_bytes::hex2bytes_unchecked(ERC20_CONTRACT_BYTECODE),
-				access_list: vec![],
-				// copied from:
-				// https://github.com/rust-ethereum/ethereum/blob/24739cc8ba6e9d8ee30ada8ec92161e4c48d578e/src/transaction.rs#L873-L875
-				odd_y_parity: false,
-				// 36b241b061a36a32ab7fe86c7aa9eb592dd59018cd0443adc0903590c16b02b0
-				r: H256([
-					54, 178, 65, 176, 97, 163, 106, 50, 171, 127, 232, 108, 122, 169, 235, 89, 45,
-					213, 144, 24, 205, 4, 67, 173, 192, 144, 53, 144, 193, 107, 2, 176,
-				]),
-				// 5edcc541b4741c5cc6dd347c5ed9577ef293a62787b4510465fadbfe39ee4094
-				s: H256([
-					54, 178, 65, 176, 97, 163, 106, 50, 171, 127, 232, 108, 122, 169, 235, 89, 45,
-					213, 144, 24, 205, 4, 67, 173, 192, 144, 53, 144, 193, 107, 2, 176,
-				]),
+				signature: TransactionSignature::new(
+					38,
+					H256([
+						190, 103, 224, 160, 125, 182, 125, 168, 212, 70, 247, 106, 221, 89, 14, 84,
+						182, 233, 44, 182, 184, 249, 131, 90, 235, 103, 84, 5, 121, 162, 119, 23,
+					]),
+					H256([
+						45, 105, 5, 22, 81, 32, 32, 23, 28, 30, 200, 112, 246, 255, 69, 57, 140,
+						200, 96, 146, 80, 50, 107, 232, 153, 21, 251, 83, 142, 123, 215, 24,
+					]),
+				)
+				.unwrap(),
 			};
-			assert_ok!(RuntimeTransact::runtime_transact(
+			assert_ok!(EthTxForwarder::runtime_transact(
 				RuntimeEthOrigin::RuntimeTransact(alice.address).into(),
-				Box::new(Transaction::EIP2930(t))
+				Box::new(Transaction::Legacy(t))
 			));
 
 			assert!(System::events()
