+++ runtime/crab/src/lib.rs
@@ -159,7 +159,7 @@ frame_support::construct_runtime! {
 		// EVM stuff.
 		Ethereum: pallet_ethereum = 36,
 		EVM: pallet_evm = 37,
-		RuntimeTransact: darwinia_runtime_transact = 38,
+		EthTxForwarder: darwinia_ethtx_forwarder = 38,
 
 		// // Crab <> Darwinia
 		// BridgePolkadotGrandpa: pallet_bridge_grandpa::<Instance1> = 39,
