+++ runtime/pangolin/Cargo.toml
@@ -44,7 +44,7 @@ darwinia-precompile-bls12-381     = { workspace = true }
 darwinia-precompile-deposit       = { workspace = true }
 darwinia-precompile-staking       = { workspace = true }
 darwinia-precompile-state-storage = { workspace = true }
-darwinia-runtime-transact         = { workspace = true }
+darwinia-ethtx-forwarder         = { workspace = true }
 darwinia-staking                  = { workspace = true }
 dc-primitives                     = { workspace = true }
 
@@ -167,7 +167,7 @@ std = [
 	"darwinia-precompile-deposit/std",
 	"darwinia-precompile-staking/std",
 	"darwinia-precompile-state-storage/std",
-	"darwinia-runtime-transact/std",
+	"darwinia-ethtx-forwarder/std",
 	"darwinia-staking/std",
 	"dc-primitives/std",
 
@@ -282,7 +282,7 @@ runtime-benchmarks = [
 	"darwinia-ecdsa-authority/runtime-benchmarks",
 	"darwinia-precompile-assets/runtime-benchmarks",
 	"darwinia-precompile-staking/runtime-benchmarks",
-	"darwinia-runtime-transact/runtime-benchmarks",
+	"darwinia-ethtx-forwarder/runtime-benchmarks",
 	"darwinia-staking/runtime-benchmarks",
 
 	# frontier
@@ -345,7 +345,7 @@ try-runtime = [
 	"darwinia-deposit/try-runtime",
 	"darwinia-ecdsa-authority/try-runtime",
 	"darwinia-message-gadget/try-runtime",
-	"darwinia-runtime-transact/try-runtime",
+	"darwinia-ethtx-forwarder/try-runtime",
 	"darwinia-staking/try-runtime",
 
 	# moonbeam
